requirements.txt
django
djangorestframework==3.12.*
psycopg2-binary>=2.8
django-environ
